requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
aleph-alpha-client
openai
google-generativeai
requests
psutil
